Account page refactored, pinned error handling
